pyproject.toml
[tool.poetry]
name = "internet-speed"
version = "0.1.0"
description = "Periodically record the internet speed and save it to a file"
authors = ["Will Jobs <github@willjobs.com>"]
license = "MIT"
readme = "README.md"
#packages = [{include = "internet_speed"}]

[tool.poetry.dependencies]
python = "^3.11"
speedtest-cli = "^2.1.3"
dropbox = "^11.36.2"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
my-script = "internet_speed.test:myfunc"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
